# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "properties": {
          "descriptor": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "场景 descriptor 对象。与 descriptor_path 二选一。",
            "title": "Descriptor"
          },
          "descriptor_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "容器内可读的 descriptor 文件路径。与 descriptor 二选一。",
            "title": "Descriptor Path"
          },
          "evaluation_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationProfilePayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "阶段二评测配置。未提供时不绑定评测模板。"
          },
          "hil_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HilConfigPayload"
              },
              {
                "type": "null"
              }
            ],
            "description": "阶段二 HIL 运行配置。未提供时保持当前纯 CARLA 控制模式。"
          }
        },
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunEnvironmentUpdateRequest": {
        "properties": {
          "debug": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug"
          },
          "weather": {
            "$ref": "#/components/schemas/WeatherPayload"
          }
        },
        "required": [
          "weather"
        ],
        "title": "RunEnvironmentUpdateRequest",
        "type": "object"
      },
...
      "RunListResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RunPayload"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务数据",
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ],
            "description": "错误信息"
          },
          "success": {
            "default": true,
            "description": "请求是否成功",
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "RunListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "这是一个用于 CARLA 0.9.16 的场景仿真控制层 MVP。\n\n- `/docs`：开发调试用的 Swagger 页面\n- `/` 或 `/ui`：最小中文 Web 控制台",
    "title": "CARLA 场景仿真控制层 MVP",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "description": "返回运行列表，可按状态过滤。",
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "description": "可选状态过滤值",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "可选状态过滤值",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询运行列表",
        "tags": [
          "运行管理"
        ]
      },
      "post": {
        "description": "根据 descriptor 创建一个 run，创建后状态为 CREATED。",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建运行",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "返回 run 当前状态、起止时间、错误原因和 artifact 路径。",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询单个运行",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/cancel": {
      "post": {
        "description": "请求取消运行。语义与 stop 类似，但标记为取消请求。",
        "operationId": "cancel_run_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取消运行",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/environment": {
      "get": {
        "description": "返回该 run 当前 descriptor 中的环境参数以及最近一次运行时环境控制值。",
        "operationId": "get_run_environment_runs__run_id__environment_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvironmentStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询运行环境控制状态",
        "tags": [
          "运行管理"
        ]
      },
      "post": {
        "description": "更新 run 的环境参数。运行中会由 executor 轮询并应用天气更新。",
        "operationId": "update_run_environment_runs__run_id__environment_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunEnvironmentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvironmentStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "更新运行环境参数",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/events": {
      "get": {
        "description": "返回该 run 的 events.jsonl 事件流。",
        "operationId": "get_run_events_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询运行事件",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/start": {
      "post": {
        "description": "将 run 放入执行队列，状态从 CREATED 变为 QUEUED。",
        "operationId": "start_run_runs__run_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "启动运行",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/stop": {
      "post": {
        "description": "请求停止运行。若尚未启动会直接取消；若运行中则进入 STOPPING。",
        "operationId": "stop_run_runs__run_id__stop_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "停止运行",
        "tags": [
          "运行管理"
        ]
      }
    },
    "/runs/{run_id}/viewer": {
      "get": {
        "description": "返回前端运行时显示界面所需的视角列表和当前可用性。",
        "operationId": "get_run_viewer_runs__run_id__viewer_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunViewerInfoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询运行时 viewer 状态",
        "tags": [
          "运行管理"
        ]
      }
    },
...
  }
}